backend/todo/requirements.txt
absl-py==1.4.0
accelerate==0.16.0
addict==2.4.0
aenum==3.1.11
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
basicsr==1.4.2
beautifulsoup4==4.11.1
bleach==6.0.0
blendmodes==2022
boltons==21.0.0
bs4==0.0.1
cachetools==5.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
clean-fid==0.1.35
click==8.1.3
coloredlogs==15.0.1
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.5
Django==4.2
djangorestframework==3.14.0
einops==0.6.0
elastic-transport==8.1.2
elasticsearch==8.3.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
facexlib==0.2.5
fastapi==0.89.0
fastjsonschema==2.16.2
ffmpy==0.3.0
filelock==3.7.1
filterpy==1.4.5
flatbuffers==23.1.21
font-roboto==0.0.1
fonts==0.0.3
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
gfpgan==1.3.8
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.16.0
google-auth-oauthlib==0.4.6
gradio==3.16.2
grpcio==1.51.1
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.12.0
humanfriendly==10.0
idna==3.3
imageio==2.25.0
inflection==0.5.1
invisible-watermark==0.1.5
ipykernel==6.21.1
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonmerge==1.9.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.5.0
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.1
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
kornia==0.6.9
lark==1.1.5
linkify-it-py==1.0.3
llvmlite==0.39.1
lmdb==1.4.0
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==2.0.5
mpmath==1.2.1
multidict==6.0.4
mypy-extensions==0.4.3
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.13.0
onnxruntime==1.13.1
opencv-contrib-python==4.7.0.68
opencv-python==4.7.0.68
openpyxl==3.0.10
orjson==3.8.5
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
piexif==1.1.3
Pillow==9.4.0
platformdirs==2.5.2
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.4
pyDeprecate==0.3.2
pydub==0.25.1
Pygments==2.14.0
pyparsing==3.0.9
pyre-extensions==0.0.23
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-json-logger==2.0.4
python-multipart==0.0.5
pytorch-lightning==1.7.7
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
realesrgan==0.3.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
resize-right==0.0.2
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rsa==4.9
safetensors==0.2.8
scikit-image==0.19.3
scipy==1.10.0
selenium==4.8.3
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
starlette==0.22.0
sympy==1.11.1
tb-nightly==2.12.0a20230126
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.17.1
tifffile==2023.2.2
timm==0.4.12
tinycss2==1.2.1
tokenizers==0.13.2
toolz==0.12.0
torch==1.13.1
torchdiffeq==0.2.3
torchmetrics==0.11.1
torchsde==0.2.5
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
trampoline==0.1.2
transformers==4.25.1
trio==0.22.0
trio-websocket==0.10.2
typing-inspect==0.8.0
typing_extensions==4.4.0
uc-micro-py==1.0.1
uri-template==1.2.0
urllib3==1.26.11
uvicorn==0.20.0
uvloop==0.17.0
virtualenv==20.16.1
watchfiles==0.18.1
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.2.2
widgetsnbextension==4.0.5
wsproto==1.2.0
xformers==0.0.16
yapf==0.32.0
yarl==1.8.2
